requirements.txt
asgiref==3.3.1
astroid==2.5.1
cffi==1.14.5
colorama==0.4.4
crispy-forms-gds==0.2.2
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
geographiclib==1.50
gunicorn==20.1.0
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.3.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.7.2
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1